feat: Agregar manejo de token en las funciones de servicio y mejorar la gestión de sesión en el store

+++ src/Services/post/deletePost.js
@@ -1,14 +1,21 @@
 const BACK_URL = import.meta.env.VITE_BACK_URL
 
+import { useUserStore } from '../../Store/useUserStore'
+
 export default async function postDelete(id) {
+  const token = useUserStore.getState().token
+  if (!token) {
+    return { status: 401, error: 'No token disponible' }
+  }
   try {
     const response = await fetch(`${BACK_URL}/api/PostsDelete/${id}`, {
       method: 'DELETE',
       headers: {
         'Content-Type': 'application/json',
         Accept: 'application/json',
+        Authorization: `Bearer ${token}`,
       },
-      credentials: 'include',
+      
     })
 
     return { status: response.status }
